+++ core/build.gradle.kts
@@ -5,6 +5,7 @@
     id(libs.plugins.kotlin.android.get().pluginId)
     id(libs.plugins.dokka.get().pluginId)
     id(libs.plugins.kotlin.parcelize.get().pluginId)
+    alias(libs.plugins.kotlin.serialization)
 }
 
 apply {
@@ -77,6 +78,7 @@ dependencies {
     implementation(libs.iban4j)
     implementation(libs.googlePlayServices.wallet)
     api(libs.gson)
+    implementation(libs.kotlinx.serializationJson)
     implementation(libs.rekisoftLazyWorker)
 
     api(libs.squareup.okhttp3.okhttp)
