+++ core/src/main/java/io/snabble/sdk/payment/PaymentCredentials.java
@@ -38,7 +38,7 @@ import io.snabble.sdk.Environment;
 import io.snabble.sdk.PaymentMethod;
 import io.snabble.sdk.R;
 import io.snabble.sdk.Snabble;
-import io.snabble.sdk.payment.externalbilling.EncryptedBillingCredentials;
+import io.snabble.sdk.payment.externalbilling.data.ExternalBillingPaymentCredentials;
 import io.snabble.sdk.payment.payone.sepa.PayoneSepaData;
 import io.snabble.sdk.utils.GsonHolder;
 import io.snabble.sdk.utils.Logger;
@@ -225,12 +225,12 @@ public class PaymentCredentials {
     /**
      * Encrypts and stores ExternalBilling payment credentials.
      */
-    public static PaymentCredentials fromExternalBilling(@NonNull final EncryptedBillingCredentials externalBilling, @NonNull final String projectId){
-        final String json = GsonHolder.get().toJson(externalBilling, EncryptedBillingCredentials.class);
+    public static PaymentCredentials fromExternalBilling(@NonNull final ExternalBillingPaymentCredentials credentials, @NonNull final String projectId){
+        final String json = GsonHolder.get().toJson(credentials, ExternalBillingPaymentCredentials.class);
         final PaymentCredentials pc = new PaymentCredentials();
         pc.generateId();
         pc.type = Type.EXTERNAL_BILLING;
-        pc.obfuscatedId = pc.obfuscate(externalBilling.getUsername());
+        pc.obfuscatedId = pc.obfuscate(credentials.getUsername());
 
         final List<X509Certificate> certificates = Snabble.getInstance().getPaymentCertificates();
         if (certificates == null || certificates.size() == 0) {
