+++ core/src/main/java/io/snabble/sdk/payment/externalbilling/data/ExternalBillingLoginResponse.kt
@@ -1,0 +1,10 @@
+package io.snabble.sdk.payment.externalbilling.data
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ExternalBillingLoginResponse(
+    @SerialName("username") val username: String,
+    @SerialName("contactPersonID") val contactPersonID: String,
+)
