+++ core/src/main/java/io/snabble/sdk/payment/externalbilling/data/ExternalBillingPaymentCredentials.kt
@@ -1,8 +1,8 @@
-package io.snabble.sdk.payment.externalbilling
+package io.snabble.sdk.payment.externalbilling.data
 
 import com.google.gson.annotations.SerializedName
 
-data class EncryptedBillingCredentials(
+data class ExternalBillingPaymentCredentials(
     @SerializedName("username") val username: String,
     @SerializedName("contactPersonID") val contactPersonID: String,
     @SerializedName("password") val password: String,
